+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/api/util/ApiServerUtils.java
@@ -22,6 +22,7 @@ import io.confluent.ksql.util.FileWatcher;
 import io.confluent.ksql.util.FileWatcher.Callback;
 import io.confluent.ksql.util.QueryMask;
 import io.confluent.ksql.util.VertxSslOptionsFactory;
+import io.netty.handler.codec.haproxy.HAProxyProtocolException;
 import io.vertx.core.http.ClientAuth;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.HttpServerOptions;
@@ -66,6 +67,8 @@ public final class ApiServerUtils {
   public static void unhandledExceptionHandler(final Throwable t) {
     if (t instanceof ClosedChannelException) {
       LOG.debug("Unhandled ClosedChannelException (connection likely closed early)", t);
+    } else if (t instanceof HAProxyProtocolException) {
+      LOG.error("Failed to decode proxy protocol header", t);
     } else {
       LOG.error("Unhandled exception", t);
     }
@@ -131,6 +134,22 @@ public final class ApiServerUtils {
   public static List<URI> parseListeners(final KsqlRestConfig config) {
     final List<String> sListeners = config.getList(KsqlRestConfig.LISTENERS_CONFIG);
     return parseListenerStrings(config, sListeners);
+  }
+
+  public static List<URI> parseProxyProtocolListeners(final KsqlRestConfig config) {
+    final List<String> sListeners = config.getList(KsqlRestConfig.PROXY_PROTOCOL_LISTENERS_CONFIG);
+    final Set<URI> listenUriSet = new HashSet<>(parseListeners(config));
+    final List<URI> proxyProtocolListenUris = parseListenerStrings(config, sListeners);
+
+    for (URI u: proxyProtocolListenUris) {
+      if (!listenUriSet.contains(u)) {
+        throw new ConfigException(String.format("Listener %s is listed in %s but not in %s.", u,
+            KsqlRestConfig.PROXY_PROTOCOL_LISTENERS_CONFIG,
+            KsqlRestConfig.LISTENERS_CONFIG));
+      }
+    }
+
+    return proxyProtocolListenUris;
   }
 
   public static void setTlsOptions(